change model to use decimal instead of double

+++ src/Trading.Bot.API/Mediator/BollingerBandsEmaHandler.cs
@@ -10,7 +10,7 @@ public class BollingerBandsEmaHandler : IRequestHandler<BollingerBandsEmaRequest
 
         var rsiHigh = request.RsiHigh ?? 70;
 
-        var maxSpread = request.MaxSpread ?? 0.0003;
+        var maxSpread = request.MaxSpread ?? 0.0003m;
 
         var minGain = request.MinGain ?? 0;
 
@@ -51,8 +51,8 @@ public record BollingerBandsEmaRequest : IHttpRequest
     public double StandardDeviation { get; set; }
     public double? RsiLow { get; set; }
     public double? RsiHigh { get; set; }
-    public double? MaxSpread { get; set; }
-    public double? MinGain { get; set; }
-    public double? RiskReward { get; set; }
+    public decimal? MaxSpread { get; set; }
+    public decimal? MinGain { get; set; }
+    public decimal? RiskReward { get; set; }
     public int? TradeRisk { get; set; }
 }
