+++ src/Trading.Bot/Models/ApiResponses/PricingResponse.cs
@@ -13,21 +13,21 @@ public class PriceResponse
     public DateTime Time { get; set; }
     public Bid[] Bids { get; set; }
     public Ask[] Asks { get; set; }
-    public double CloseoutBid { get; set; }
-    public double CloseoutAsk { get; set; }
+    public decimal CloseoutBid { get; set; }
+    public decimal CloseoutAsk { get; set; }
     public string Status { get; set; }
     public bool Tradeable { get; set; }
     public string Instrument { get; set; }
 
     public class Bid
     {
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int Liquidity { get; set; }
     }
 
     public class Ask
     {
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int Liquidity { get; set; }
     }
 }
@@ -35,7 +35,7 @@ public class PriceResponse
 public class HomeConversionResponse
 {
     public string Currency { get; set; }
-    public double AccountGain { get; set; }
-    public double AccountLoss { get; set; }
-    public double PositionValue { get; set; }
+    public decimal AccountGain { get; set; }
+    public decimal AccountLoss { get; set; }
+    public decimal PositionValue { get; set; }
 }
